feat: Create migrations

+++ src/Domain/Entities/OrderItem.cs
@@ -5,9 +5,9 @@ namespace Domain.Entities
         public long Id { get; private set; }
         public long OrderId { get; private set; }
         public long ItemId { get; private set; }
-        public long Quantity { get; private set; }
+        public int Quantity { get; private set; }
 
-        public OrderItem(long orderId, long itemId, long quantity)
+        public OrderItem(long orderId, long itemId, int quantity)
         {
             OrderId = orderId;
             ItemId = itemId;
